Project

+++ MyComplexApp/App.Services.Dtos/DTOs/ThemeMark/AddThemeMarkDto.cs
@@ -6,8 +6,7 @@ namespace App.Services.Dtos.DTOs.ThemeMark
 {
     public class AddThemeMarkDto
     {
-        public int Mark { get; set; }
-
+        public double Mark { get; set; }
         public string Comment { get; set; }
         public long? InternId { get; set; }
         public long? ThemeId { get; set; }
